Desktop/Syntheverse-Holographic-RAG/requirements_langchain.txt
# LangChain-based RAG Pipeline Requirements
# Based on LangChain (121K+ stars on GitHub) - the most popular RAG framework

# Core LangChain packages
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5

# Document loaders and processing
pypdf>=3.17.0
pdfplumber>=0.10.0
unstructured>=0.11.0  # Optional, for advanced PDF processing

# Vector store
chromadb>=0.4.22

# OpenAI API
openai>=1.10.0

# Web scraping and HTTP
requests>=2.31.0
beautifulsoup4>=4.12.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.5.0

# Optional: For better PDF processing
# unstructured[pdf]  # Uncomment if you want to use UnstructuredPDFLoader
